wishlist add

+++ src/utils/utils.js
@@ -1,3 +1,5 @@
+import toast from "react-hot-toast";
+
 const get = () => {
   const gadgets = localStorage.getItem("favorites");
   if (gadgets) {
@@ -7,9 +9,31 @@ const get = () => {
     return [];
   }
 };
+const getWL = () => {
+  const gadgets = localStorage.getItem("wishlist");
+  if (gadgets) {
+    const data = JSON.parse(gadgets);
+    return data;
+  } else {
+    return [];
+  }
+};
 
-export const add = (gadget) => {
+const add = (gadget) => {
   const favorites = get();
+  const exist = favorites.find((data) => data.product_id === gadget.product_id);
+  if (exist) return toast.error("Item already exist");
   favorites.push(gadget);
+  toast.success("Item added successfull in the Cart");
   localStorage.setItem("favorites", JSON.stringify(favorites));
 };
+const addwl = (gadget) => {
+  const wishlist = getWL();
+  const exist = wishlist.find((data) => data.product_id === gadget.product_id);
+  if (exist) return toast.error("Item already exist");
+  wishlist.push(gadget);
+  toast.success("Item added successfull in the Wishlist");
+  localStorage.setItem("wishlist", JSON.stringify(wishlist));
+};
+
+export { add, get, addwl, getWL };
